refactor: separate tsping to separate subpackage
