+++ src/keri/spac/essr/__init__.py
@@ -1,0 +1,6 @@
+# -*- encoding: utf-8 -*-
+"""
+KERI
+keri.spac.essr Package
+"""
+
