+++ tests/app/test_forwarding.py
@@ -16,6 +16,7 @@ from keri.core import coring, eventing, parsing, serdering
 from keri.app import forwarding, habbing, indirecting, storing
 
 from keri.peer import exchanging
+from keri.spac.essr import tsping
 
 
 def test_postman(seeder):
@@ -98,13 +99,13 @@ def test_essr_stream(seeder):
         httpServerDoer = http.ServerDoer(server=server)
 
         kvy = eventing.Kevery(db=hab.db)
-        parsing.Parser(version=Vrsn_1_0).parseOne(bytearray(recpHab.makeOwnEvent(sn=0)), kvy=kvy, local=True)
+        parsing.Parser.parseOne(bytearray(recpHab.makeOwnEvent(sn=0)), kvy=kvy, local=True)
         kvy.processEscrows()
         assert recpHab.pre in kvy.kevers
 
         recpKvy = eventing.Kevery(db=recpHab.db)
         icp = hab.makeOwnEvent(sn=0)
-        parsing.Parser(version=Vrsn_1_0).parseOne(bytearray(icp), kvy=recpKvy, local=True)
+        parsing.Parser.parseOne(bytearray(icp), kvy=recpKvy, local=True)
         kvy.processEscrows()
         assert hab.pre in recpKvy.kevers
 
@@ -160,8 +161,8 @@ def test_essr_stream(seeder):
         texter = recpHby.db.essrs.get(essrSaidA)[0]
         ims = bytearray(recpHab.decrypt(texter.raw))
 
-        tag = recpHby.psr.extract(ims, coring.Tsper)
-        assert tag.tsp == coring.Tsps.SCS
+        tag = recpHby.psr.extract(ims, tsping.Tsper)
+        assert tag.tsp == tsping.Tsps.SCS
         pre = recpHby.psr.extract(ims, coring.Prefixer)
         assert pre.qb64 == hab.pre  # encrypt sender
         pad = recpHby.psr.extract(ims, coring.Bexter)
@@ -171,7 +172,7 @@ def test_essr_stream(seeder):
         texter = recpHby.db.essrs.get(essrSaidB)[0]
         ims = bytearray(recpHab.decrypt(texter.raw))
 
-        _tag = recpHby.psr.extract(ims, coring.Tsper)
+        _tag = recpHby.psr.extract(ims, tsping.Tsper)
         _pre = recpHby.psr.extract(ims, coring.Prefixer)
         _pad = recpHby.psr.extract(ims, coring.Bexter)
         recpHby.psr.parseOne(ims=ims)
@@ -263,8 +264,8 @@ def test_essr_mbx(seeder):
         texter = wesHby.db.essrs.get(essrSaidA)[0]
         ims = bytearray(wesHab.decrypt(texter.raw))
 
-        tag = wesHby.psr.extract(ims, coring.Tsper)
-        assert tag.tsp == coring.Tsps.SCS
+        tag = wesHby.psr.extract(ims, tsping.Tsper)
+        assert tag.tsp == tsping.Tsps.SCS
         pre = wesHby.psr.extract(ims, coring.Prefixer)
         assert pre.qb64 == hab.pre  # encrypt sender
         pad = wesHby.psr.extract(ims, coring.Bexter)
@@ -281,7 +282,7 @@ def test_essr_mbx(seeder):
         texter = wesHby.db.essrs.get(essrSaidB)[0]
         ims = bytearray(wesHab.decrypt(texter.raw))
 
-        _tag = wesHby.psr.extract(ims, coring.Tsper)
+        _tag = wesHby.psr.extract(ims, tsping.Tsper)
         _pre = wesHby.psr.extract(ims, coring.Prefixer)
         _pad = wesHby.psr.extract(ims, coring.Bexter)
         parser.parse(ims=ims)
