+++ tests/spac/essr/__init__.py
@@ -1,0 +1,5 @@
+# -*- encoding: utf-8 -*-
+"""
+tests.spac.essr module
+
+"""
